move tests from ingest to rpc service
